# testing/clase_2/clase_2_TM/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Goesgil",
            "url": "http://www.example.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/transactions": {
            "post": {
                "security": [
                    {
                        "token header string \"Bearer \u003ctoken\u003e\"": []
                    }
                ],
                "description": "Create a new transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trxs"
                ],
                "summary": "Transactions",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/web.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "put": {
                "security": [
                    {
                        "token header string \"Bearer \u003ctoken\u003e\"": []
                    }
                ],
                "description": "Update a transaction",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "trxs"
                ],
                "summary": "Transactions",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/web.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "token header string \"Bearer \u003ctoken\u003e\"": []
                    }
                ],
                "description": "Update codeTrxs and amount of a transaction",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "trxs"
                ],
                "summary": "Transactions",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/web.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "token header string \"Bearer \u003ctoken\u003e\"": []
                    }
                ],
                "description": "Update codeTrxs and amount of a transaction",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "trxs"
                ],
                "summary": "Transactions",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/web.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "web.response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        }
    }
}